requirements.txt
Django==5.2.12
djangorestframework
drf-spectacular
pandas